# Remove most calls to time.Now