@mrd0ll4r 2017-05-17
Stats still work even without timing them :)